extract persist
